database_requirements.txt
# Database setup script requirements
psycopg2-binary==2.9.9
python-dotenv==1.0.0